refactor(web): make the no-cycle guard functional; break intra-core cycles
Review of the initial P3a guard found it inert: import-x/no-cycle had no
parser or resolver wired, so it couldn't resolve the @/* alias every
data-layer edge uses and silently traversed nothing — passing green while
a cycle existed. Wire the TypeScript parser + resolver and add
no-unresolved on the same scope as a loud tripwire, verified against an
injected fixture cycle that now errors.

Turning the guard on exposed 3 pre-existing intra-hooks/github runtime
cycles (mutations<->queries and rulesets->queries->mutations) via shared
primitives. Extract those primitives — paginateAll, getRepo, createTeam,
getErrorMessage, the collect/regrade workflow constants, and the
GitHubTreeResponse type — into dependency-light github-core leaf modules
(paginate, repoReads, teamWrites, errorMessage, workflows; the type into
types.ts) that both files import downward. The data layer is now
cycle-free under the guard. The logger<->activityStore<->errors init
cycle stays deliberately mitigated (lazy logger) with a scoped disable.

Also: move the stale UploadRoster test mock onto @/api/mutations/students
(the barrel path it mocked is gone) and trim the configRepoReads comment.
