+++ web/src/hooks/github/errorMessage.ts
@@ -1,0 +1,4 @@
+export function getErrorMessage(error: unknown): string {
+  if (error instanceof Error) return error.message
+  return "Unknown GitHub API error"
+}
