+++ web/src/hooks/github/queries.ts
@@ -13,6 +13,7 @@ import type {
   GitHubRelease,
   GitHubRepo,
   GitHubTeam,
+  GitHubTreeResponse,
   GitHubUser,
   GitHubWorkflowRun,
 } from "./types"
@@ -25,13 +26,10 @@ import {
   retryTransientGitHubError,
   tolerateGitHubError,
 } from "./errors"
-import {
-  COLLECT_SCORES_WORKFLOW,
-  REGRADE_WORKFLOW,
-  createTeam,
-  getErrorMessage,
-  type GitHubTreeResponse,
-} from "./mutations"
+import { COLLECT_SCORES_WORKFLOW, REGRADE_WORKFLOW } from "./workflows"
+import { createTeam } from "./teamWrites"
+import { getErrorMessage } from "./errorMessage"
+import { paginateAll } from "./paginate"
 import { decodeBase64Utf8 } from "@/util/github"
 import { mapWithConcurrency } from "@/util/concurrency"
 import { getCommit } from "./configRepoReads"
@@ -849,34 +847,6 @@ export async function listClassroomDirs(
   )
 }
 
-// Walk a GitHub list endpoint to exhaustion, 100 items per page. `makePath`
-// receives the 1-based page number. Stops when a page returns fewer than 100.
-export async function paginateAll<T>(
-  client: GitHubClient,
-  makePath: (page: number) => string,
-): Promise<T[]> {
-  const all: T[] = []
-  let page = 1
-  // Hard cap (100 pages x 100/page = 10k items) so a server that ignores the
-  // page param and keeps returning full pages can't loop unbounded.
-  const MAX_PAGES = 100
-
-  while (page <= MAX_PAGES) {
-    const batch = await client.request<T[]>(makePath(page))
-    all.push(...batch)
-    if (batch.length < 100) break
-    page++
-  }
-
-  if (page > MAX_PAGES) {
-    log.warn("pagination hit MAX_PAGES cap, results may be truncated", {
-      maxPages: MAX_PAGES,
-    })
-  }
-
-  return all
-}
-
 // Failed / expired org invitations (carry failed_at / failed_reason). Owner-only.
 // Read org-wide, then attributed to a classroom team by
 // getOrgFailedInvitationsForTeam (GitHub has no team-scoped failed endpoint).
@@ -1195,17 +1165,6 @@ export async function getClassroom50OrgSummary(
 // into hundreds of concurrent requests (GitHub secondary-rate-limit territory)
 // while still beating a strictly-sequential loop.
 export const REPO_READ_CONCURRENCY = 8
-
-export async function getRepo(
-  client: GitHubClient,
-  owner: string,
-  repo: string,
-) {
-  return tolerateGitHubError(
-    () => client.request<GitHubRepo>(`/repos/${owner}/${repo}`),
-    null,
-  )
-}
 
 // List a team's members across all pages. 404 (team doesn't exist yet) returns
 // [] so a classroom whose staff team hasn't been created reads as "no members".
