+++ web/src/hooks/useTriggerRegrade.ts
@@ -4,7 +4,8 @@ import { useTranslation } from "react-i18next"
 import { useGitHubClient } from "@/context/github/GitHubProvider"
 import { useActionActivityRegistry } from "@/context/actions/ActionActivityProvider"
 import { useRegradeCoordinator } from "@/context/regrade/RegradeCoordinator"
-import { REGRADE_WORKFLOW, triggerRegrade } from "./github/mutations"
+import { triggerRegrade } from "./github/mutations"
+import { REGRADE_WORKFLOW } from "./github/workflows"
 import { getRegradeRunAfterId, githubKeys } from "./github/queries"
 import { useGitHubOperation, type OperationPhase } from "./useGitHubOperation"
