+++ web/src/pages/SubmissionsPage.tsx
@@ -63,7 +63,7 @@ import RoleResolvingFallback from "@/components/RoleResolvingFallback"
 import {
   COLLECT_SCORES_WORKFLOW,
   REGRADE_WORKFLOW,
-} from "@/hooks/github/mutations"
+} from "@/hooks/github/workflows"
 import {
   formatDueDateTime,
   formatRelativeToNow,
