+++ web/src/pages/orgMembers/removeMemberFromOrg.ts
@@ -1,7 +1,8 @@
 import type { GitHubClient } from "@/hooks/github/client"
 import type { TFunction } from "i18next"
 import { unenrollStudent } from "@/api/mutations/students"
-import { removeOrgMembership, getErrorMessage } from "@/hooks/github/mutations"
+import { removeOrgMembership } from "@/hooks/github/mutations"
+import { getErrorMessage } from "@/hooks/github/errorMessage"
 import { getAuthenticatedUser } from "@/api/queries/users"
 import { isSameGitHubUser } from "@/util/students"
 import type { Student } from "@/types/classroom"
