+++ web/src/pages/students/bulkUnenrollRoster.test.ts
@@ -12,7 +12,7 @@ const bulkUnenrollMock = vi.fn()
 vi.mock("@/api/mutations/students", () => ({
   bulkUnenrollStudents: (...args: unknown[]) => bulkUnenrollMock(...args),
 }))
-vi.mock("@/hooks/github/mutations", () => ({
+vi.mock("@/hooks/github/errorMessage", () => ({
   getErrorMessage: (err: unknown) =>
     err instanceof Error ? err.message : String(err),
 }))
